Added downstream API call for Category, Catalog, ProductOfferingPrice, ProductOffering, ProductSpecification
